agents/adk-wikipedia/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pyyaml==6.0.3
requests==2.32.5
google-adk==1.15.1
wikipedia==1.4.0